+++ models/userpoints.js
@@ -16,7 +16,7 @@ module.exports = (sequelize, DataTypes) => {
   UserPoints.init({
     userId: DataTypes.INTEGER,
     quizzId: DataTypes.INTEGER,
-    point:DataTypes.INTEGER,
+    point:DataTypes.DECIMAL,
     correctAnswers:DataTypes.INTEGER,
     courseId:DataTypes.INTEGER,
     isFinal:DataTypes.BOOLEAN
